+++ tests/google_big_query/test_google_big_query.py
@@ -1,13 +1,15 @@
-from typing import Any, Generator
+from typing import Any, Callable, Generator
 from unittest.mock import patch
 
 import pandas
 import pandas as pd
 import pytest
 from google.api_core.exceptions import NotFound
+from google.cloud import bigquery
 from google.cloud.bigquery import ArrayQueryParameter, Client, ScalarQueryParameter
 from google.cloud.bigquery.job.query import QueryJob
 from google.cloud.bigquery.table import RowIterator
+from google.cloud.exceptions import Unauthorized
 from google.oauth2.service_account import Credentials
 from pandas.util.testing import assert_frame_equal  # <-- for testing dataframes
 from pytest_mock import MockFixture
@@ -15,13 +17,16 @@ from pytest_mock import MockFixture
 from toucan_connectors.google_big_query.google_big_query_connector import (
     GoogleBigQueryConnector,
     GoogleBigQueryDataSource,
+    InvalidJWTToken,
     _define_query_param,
 )
 from toucan_connectors.google_credentials import GoogleCredentials
 
+import_path = 'toucan_connectors.google_big_query.google_big_query_connector'
+
 
 @pytest.fixture
-def _fixture_credentials():
+def _fixture_credentials() -> GoogleCredentials:
     my_credentials = GoogleCredentials(
         type='my_type',
         project_id='my_project_id',
@@ -35,6 +40,16 @@ def _fixture_credentials():
         client_x509_cert_url='https://www.googleapis.com/robot/v1/metadata/x509/pika.com',
     )
     return my_credentials
+
+
+@pytest.fixture
+def gbq_connector_with_jwt(_fixture_credentials: GoogleCredentials) -> GoogleBigQueryConnector:
+    return GoogleBigQueryConnector(
+        name='woups',
+        scopes=['https://www.googleapis.com/auth/bigquery'],
+        credentials=_fixture_credentials,
+        jwt_token='this-is-a-jwt-token',
+    )
 
 
 @pytest.fixture
@@ -121,14 +136,50 @@ def test_connect(load_pem_private_key, client, _fixture_credentials, _fixture_sc
     assert isinstance(connection, Client)
 
 
+def test__http_is_present_as_attr(
+    mocker: MockFixture,
+    gbq_connector_with_jwt: GoogleBigQueryConnector,
+) -> None:
+    """we should have _http as arg to bigquery.Client when the jwt is provided in google-credentials"""
+    mock_bigqueryClient = mocker.patch('google.cloud.bigquery.Client')
+    gbq_connector_with_jwt._get_bigquery_client()
+    assert mock_bigqueryClient.call_count == 1
+    # we ensure that _http is inside the list of called args
+    assert ['project', '_http'] == list(mock_bigqueryClient.call_args[1].keys())
+
+
+def test_http_connect(
+    mocker: MockFixture,
+    gbq_connector_with_jwt: GoogleBigQueryConnector,
+) -> None:
+    """we should call for _http_connect when the jwt is provided in google-credentials"""
+    mock_http_connect = mocker.patch(f'{import_path}.GoogleBigQueryConnector._http_connect')
+    gbq_connector_with_jwt._get_bigquery_client()
+    assert mock_http_connect.call_count == 1
+    assert ['http_session', 'project_id'] == list(mock_http_connect.call_args[1].keys())
+
+
+def test_http_connect_on_invalid_token(
+    mocker: MockFixture,
+    gbq_connector_with_jwt: GoogleBigQueryConnector,
+) -> None:
+    """we should have _http as arg to bigquery.Client when the jwt is provided in google-credentials"""
+    mock_bigqueryClient = mocker.patch('google.cloud.bigquery.Client')
+    mock_bigqueryClient.side_effect = Unauthorized('Error with the JWT token')
+    with pytest.raises(InvalidJWTToken):
+        gbq_connector_with_jwt._get_bigquery_client()
+
+
 @patch(
-    'google.cloud.bigquery.table.RowIterator.to_dataframe',
-    return_value=pandas.DataFrame({'a': [1, 1], 'b': [2, 2]}),
+    'google.cloud.bigquery.table.RowIterator.to_dataframe_iterable',
+    return_value=iter((pandas.DataFrame({'a': [1, 1], 'b': [2, 2]}),)),
 )
 @patch('google.cloud.bigquery.job.query.QueryJob.result', return_value=RowIterator)
 @patch('google.cloud.bigquery.Client.query', return_value=QueryJob)
 @patch('google.cloud.bigquery.Client', autospec=True)
-def test_execute(client, execute, result, to_dataframe):
+def test_execute(
+    client: bigquery.Client, execute: Callable, result: pd.DataFrame, to_dataframe: Callable
+):
     result = GoogleBigQueryConnector._execute_query(client, 'SELECT 1 FROM my_table', [])
     assert_frame_equal(pandas.DataFrame({'a': [1, 1], 'b': [2, 2]}), result)
 
@@ -685,3 +736,41 @@ def test_get_form(mocker: MockFixture, _fixture_credentials: MockFixture) -> Non
         )['properties']['database']['default']
         == 'my_project_id'
     )
+
+
+@pytest.mark.parametrize(
+    'input_query, expected_output',
+    [
+        # Test cases with double quotes (unchanged)
+        ('SELECT "column1" FROM table', 'SELECT `column1` FROM table'),
+        ('"quoted text" inside query', '`quoted text` inside query'),
+        # Test cases with '@__' and '__''
+        ('@__param1__ in query', '@__param1__ in query'),  # No change for param1
+        (
+            '"quoted text"@__param2__',
+            '`quoted text`@__param2__',
+        ),  # Replace double quotes, keep param2 intact
+        (
+            '@__param3__ and @__param4__',
+            '@__param3__ and @__param4__',
+        ),  # No change for params 3 and 4
+        # Test cases with escaped single quotes
+        ("'single-quoted text'", "'single-quoted text'"),  # No change for single-quoted text
+        (
+            "'escaped \\' single quote'",
+            "'escaped \\' single quote'",
+        ),  # No change for escaped single quote
+        ("'@__param7__'", '@__param7__'),  # Remove surrounding single quotes for param7
+        (
+            "'@__param8__' and '@__param9__'",
+            '@__param8__ and @__param9__',
+        ),  # Remove surrounding single quotes for params 8 and 9
+        # Mixed cases
+        (
+            '@__param5__ "with quotes" @__param6__',
+            '@__param5__ `with quotes` @__param6__',
+        ),  # No change for params 5 and 6
+    ],
+)
+def test_clean_query(input_query, expected_output):
+    assert GoogleBigQueryConnector._clean_query(input_query) == expected_output