+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+## Changed
+
+- Goole Big Query: Now support signed JWT connection on the GBQ connector.
+
 ### [4.8.1] 2023-09-18
 
 ### Fixed